+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["react-native", "module:react-native-dotenv"]
+  "presets": ["react-native"]
 }
