+++ index.js
@@ -1,1 +1,7 @@
-import "./src/app/main";
+/** @format */
+
+import { AppRegistry } from "react-native";
+import App from "./src/app/App";
+import { name as appName } from "./src/app/app.json";
+
+AppRegistry.registerComponent(appName, () => App);
