+++ package.json
@@ -9,11 +9,9 @@
   "dependencies": {
     "react": "16.4.1",
     "react-native": "0.56.0",
-    "react-native-dotenv": "^0.2.0",
-    "react-native-vector-icons": "^5.0.0",
-    "react-navigation": "^2.9.3",
-    "undefined": "^0.1.0",
-    "unstated": "^2.1.1"
+    "react-native-svg": "^6.4.1",
+    "react-navigation": "^2.6.2",
+    "undefined": "^0.1.0"
   },
   "devDependencies": {
     "babel-jest": "23.4.0",
