--- src/app/main.js
@@ -1,5 +1,0 @@
-import { AppRegistry } from "react-native";
-import { RootComponent } from "./RootComponent";
-import { name as appName } from "./app.json";
-
-AppRegistry.registerComponent(appName, () => RootComponent);
